Added class docstring

+++ src/torchmetrics/functional/regression/minkowski.py
@@ -7,7 +7,7 @@ from torchmetrics.utilities.checks import _check_same_shape
 from torchmetrics.utilities.exceptions import TorchMetricsUserError
 
 
-def _minkowski_distance_update(preds: Tensor, targets: Tensor, p: int) -> Tensor:
+def _minkowski_distance_update(preds: Tensor, targets: Tensor, p: float) -> Tensor:
     _check_same_shape(preds, targets)
 
     if p < 0:
@@ -19,10 +19,10 @@ def _minkowski_distance_update(preds: Tensor, targets: Tensor, p: int) -> Tensor
     return mink_dist_sum
 
 
-def _minkowski_distance_compute(distance: Tensor, p: int) -> Tensor:
+def _minkowski_distance_compute(distance: Tensor, p: float) -> Tensor:
     return torch.pow(distance, 1.0 / p)
 
 
-def minkowski_distance(preds: Tensor, targets: Tensor, p: int) -> Tensor:
+def minkowski_distance(preds: Tensor, targets: Tensor, p: float) -> Tensor:
     minkowski_dist_sum = _minkowski_distance_update(preds, targets, p)
     return _minkowski_distance_compute(minkowski_dist_sum, p)
